Better system info output for OS other than Linux. Temporarily disabled manual mipmap generation in OpenGL as it is wrong code.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@185 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CSceneNodeAnimatorRotation.cpp
@@ -34,7 +34,7 @@ void CSceneNodeAnimatorRotation::animateNode(ISceneNode* node, u32 timeMs)
 		if ((timeMs-StartTime) != 0)
 			DiffTime=timeMs-StartTime;
 		core::vector3df NewRotation = node->getRotation();
-		NewRotation += Rotation* (DiffTime/10.0f);
+		NewRotation += Rotation* ((f32)DiffTime/10.0f);
 		node->setRotation(NewRotation);
 		StartTime=timeMs;
 	}
